+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Joyride
 
 ## [Unreleased]
 
+- [Make it easy to open scripts for editing](https://github.com/BetterThanTomorrow/joyride/issues/56)
+
 ## [0.0.9] - 2022-05-11
 - [Error when dismissing the scripts menu without selecting anything](https://github.com/BetterThanTomorrow/joyride/issues/24)
 - Fix: [Better error report when a script isn't found](https://github.com/BetterThanTomorrow/joyride/issues/40)